+++ CHANGELOG.md
@@ -9,7 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-* None yet!
+### Updated
+
+* Logging was refactored to use app settings
 
 ## [0.23.1] - 09/29/2023